feat(sequence)!: extract ironrdp-sequence crate
Create ironrdp-sequence, a new sans-I/O core-tier crate that owns the
`Sequence` trait and its supporting contract types previously defined
in ironrdp-connector: `State`, `Written`, `SequenceError`/
`SequenceErrorKind`/`SequenceResult`, the `general_err!`/`reason_err!`/
`custom_err!` macros, `ServerName`, and `DesktopSize`.

Also move `MonotonicInstant` out of ironrdp-core into ironrdp-sequence,
alongside `Sequence`. Per the PR #1530 audit, the clock read by a
sequence's `step` is part of the sans-I/O contract, not a foundational
encoding primitive, so it no longer belongs in ironrdp-core.

`MonotonicInstant` and `DesktopSize` need no allocation and stay
available unconditionally. Everything else needs `ironrdp-pdu`, so it
is gated behind a default-disabled `state-machine` feature: this keeps
ironrdp-rdpeudp, which depends on this crate only for
`MonotonicInstant`, from acquiring `ironrdp-pdu`'s much larger
dependency tree. A new xtask dependency-edge check enforces this.

ironrdp-connector and ironrdp-rdpeudp both keep permanent facade
re-exports of everything they used to define directly, so existing
`ironrdp_connector::{Sequence, MonotonicInstant, ...}` and
`ironrdp_rdpeudp::MonotonicInstant` paths are unaffected.

Stacked on #1751 (branch cbenoit-introduce-sequenceerror-sspi-free-sequen).
Closes #1426 once this stack lands on master.

BREAKING CHANGE: `ironrdp_core::MonotonicInstant` is removed. Use
`ironrdp_sequence::MonotonicInstant`, or the stable facade re-exports
at `ironrdp_connector::MonotonicInstant` / `ironrdp_rdpeudp::MonotonicInstant`.
