+++ crates/ironrdp-connector/Cargo.toml
@@ -26,6 +26,7 @@ ironrdp-svc = { path = "../ironrdp-svc", version = "0.8" } # public
 ironrdp-core = { path = "../ironrdp-core", version = "0.2" } # public
 ironrdp-error = { path = "../ironrdp-error", version = "0.2" } # public
 ironrdp-pdu = { path = "../ironrdp-pdu", version = "0.9", features = ["std"] } # public
+ironrdp-sequence = { path = "../ironrdp-sequence", version = "0.1", features = ["state-machine", "std"] } # public
 sspi = { version = "0.21", features = ["scard"] }
 url = "2.5" # public
 rand = { version = "0.9", features = ["std"] } # TODO: dependency injection?
