+++ crates/ironrdp-connector/src/lib.rs
@@ -1,18 +1,13 @@
 #![cfg_attr(doc, doc = include_str!("../README.md"))]
 #![doc(html_logo_url = "https://cdnweb.devolutions.net/images/projects/devolutions/logos/devolutions-icon-shadow.svg")]
 
-mod macros;
-
 mod channel_connection;
 mod connection;
 pub mod connection_activation;
 mod connection_finalization;
 pub mod credssp;
 mod license_exchange;
-mod sequence_error;
-mod server_name;
 
-use core::any::Any;
 use core::fmt;
 use std::sync::Arc;
 
@@ -21,7 +16,7 @@ use ironrdp_pdu::nego::NegoRequestData;
 use ironrdp_pdu::rdp::capability_sets::{self, BitmapCodecs};
 use ironrdp_pdu::rdp::client_info::{self, PerformanceFlags, TimezoneInfo};
 use ironrdp_pdu::x224::X224;
-use ironrdp_pdu::{PduHint, gcc, x224};
+use ironrdp_pdu::{gcc, x224};
 pub use sspi;
 
 pub use self::channel_connection::{ChannelConnectionSequence, ChannelConnectionState};
@@ -31,69 +26,23 @@ pub use self::connection::{
 };
 pub use self::connection_finalization::{ConnectionFinalizationSequence, ConnectionFinalizationState};
 pub use self::license_exchange::{LicenseExchangeSequence, LicenseExchangeState};
-pub use self::sequence_error::{SequenceError, SequenceErrorExt, SequenceErrorKind, SequenceResult, SequenceResultExt};
-pub use self::server_name::ServerName;
 pub use crate::license_exchange::LicenseCache;
 /// Re-exported so `connect_*`/`accept_*` boundary functions across
 /// `ironrdp-async`, `ironrdp-blocking`, and `ironrdp-acceptor` can call
 /// `.map_err_as::<ConnectorErrorKind>()` without a direct `ironrdp-error`
 /// dependency. See the [`ironrdp_error::ErrorMapping`] impl on [`ConnectorErrorKind`] below.
 pub use ironrdp_error::ResultExt;
-
-/// Provides user-friendly error messages for RDP negotiation failures
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct NegotiationFailure(ironrdp_pdu::nego::FailureCode);
-
-impl NegotiationFailure {
-    pub fn code(self) -> ironrdp_pdu::nego::FailureCode {
-        self.0
-    }
-}
-
-impl core::error::Error for NegotiationFailure {}
-
-impl From<ironrdp_pdu::nego::FailureCode> for NegotiationFailure {
-    fn from(code: ironrdp_pdu::nego::FailureCode) -> Self {
-        Self(code)
-    }
-}
-
-impl fmt::Display for NegotiationFailure {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use ironrdp_pdu::nego::FailureCode;
-
-        match self.0 {
-            FailureCode::SSL_REQUIRED_BY_SERVER => {
-                write!(f, "server requires Enhanced RDP Security with TLS or CredSSP")
-            }
-            FailureCode::SSL_NOT_ALLOWED_BY_SERVER => {
-                write!(f, "server only supports Standard RDP Security")
-            }
-            FailureCode::SSL_CERT_NOT_ON_SERVER => {
-                write!(f, "server lacks valid authentication certificate")
-            }
-            FailureCode::INCONSISTENT_FLAGS => {
-                write!(f, "inconsistent security protocol flags")
-            }
-            FailureCode::HYBRID_REQUIRED_BY_SERVER => {
-                write!(f, "server requires Enhanced RDP Security with CredSSP")
-            }
-            FailureCode::SSL_WITH_USER_AUTH_REQUIRED_BY_SERVER => {
-                write!(
-                    f,
-                    "server requires Enhanced RDP Security with TLS and client certificate"
-                )
-            }
-            _ => write!(f, "unknown negotiation failure (code: 0x{:08x})", u32::from(self.0)),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct DesktopSize {
-    pub width: u16,
-    pub height: u16,
-}
+/// Permanent facade re-export. `ironrdp-sequence` is the sans-I/O state-machine
+/// contract crate that these items were extracted into; this crate keeps
+/// re-exporting them so existing `ironrdp_connector::{...}` import paths remain
+/// stable.
+pub use ironrdp_sequence::{
+    DesktopSize, MonotonicInstant, NegotiationFailure, Sequence, SequenceError, SequenceErrorExt, SequenceErrorKind,
+    SequenceResult, SequenceResultExt, ServerName, State, Written, state_downcast, state_is,
+};
+/// Permanent facade re-export of the macros previously defined directly in this
+/// crate; see [`ironrdp_sequence`] for their definitions.
+pub use ironrdp_sequence::{custom_err, general_err, reason_err};
 
 #[derive(Debug, Clone)]
 pub struct BitmapConfig {
@@ -302,116 +251,6 @@ pub struct Config {
 }
 
 ironrdp_core::assert_impl!(Config: Send, Sync);
-
-pub trait State: Send + fmt::Debug + 'static {
-    fn name(&self) -> &'static str;
-    fn is_terminal(&self) -> bool;
-    fn as_any(&self) -> &dyn Any;
-}
-
-ironrdp_core::assert_obj_safe!(State);
-
-pub fn state_downcast<T: State>(state: &dyn State) -> Option<&T> {
-    state.as_any().downcast_ref()
-}
-
-pub fn state_is<T: State>(state: &dyn State) -> bool {
-    state.as_any().is::<T>()
-}
-
-impl State for () {
-    fn name(&self) -> &'static str {
-        "()"
-    }
-
-    fn is_terminal(&self) -> bool {
-        true
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum Written {
-    Nothing,
-    Size(core::num::NonZeroUsize),
-}
-
-impl Written {
-    #[inline]
-    pub fn from_size(value: usize) -> SequenceResult<Self> {
-        core::num::NonZeroUsize::new(value)
-            .map(Self::Size)
-            .ok_or_else(|| SequenceError::general("invalid written length (can't be zero)"))
-    }
-
-    #[inline]
-    pub fn is_nothing(self) -> bool {
-        matches!(self, Self::Nothing)
-    }
-
-    #[inline]
-    pub fn size(self) -> Option<usize> {
-        if let Self::Size(size) = self {
-            Some(size.get())
-        } else {
-            None
-        }
-    }
-}
-
-/// A point on a monotonic millisecond clock owned by the I/O driver.
-///
-/// Lives in `ironrdp-core`, shared with `ironrdp-rdpeudp`, and re-exported
-/// here. The epoch is arbitrary and carries no meaning; only differences
-/// between two instants do. All instants passed to one connector across its
-/// lifetime must come from the same clock: comparing instants from two
-/// different epochs produces a meaningless delta rather than an error, either
-/// saturating to zero or landing on a huge value with no diagnostic.
-///
-/// The clock deliberately lives outside the sans-I/O sequences, because a
-/// sequence reading a clock itself would measure how quickly it drained an
-/// already-filled buffer rather than how long the bytes took to arrive. Only
-/// the driver that performed the read knows the latter, which is what the
-/// `None` in [`Sequence::step`] is for: a driver with no reading to pass on.
-/// `ironrdp-blocking` and `ironrdp-async` each stamp with their own
-/// driver-owned epoch; the FFI connector, which has no read loop of its own to
-/// time, stamps on entry to its `step` binding instead.
-pub use ironrdp_core::MonotonicInstant;
-
-pub trait Sequence: Send {
-    fn next_pdu_hint(&self) -> Option<&dyn PduHint>;
-
-    fn state(&self) -> &dyn State;
-
-    /// Advances the sequence.
-    ///
-    /// `received_at` is when `input` arrived on the wire, as observed by the I/O
-    /// driver, or `None` from a driver that does not observe arrival times. The
-    /// absence of a reading is deliberately not expressible as an instant: a
-    /// driver that cannot measure has taken no measurement, which is a different
-    /// thing from one that measured no elapsed time, and only the sequence
-    /// knows which of the two its reply may be derived from.
-    ///
-    /// A driver that always passes `None` never opens a connect-time bandwidth
-    /// window, so the Bandwidth Measure Results it sends report only the Stop's
-    /// own payload against the untimed floor. See `connection::counted_len`'s doc
-    /// for why the byte count is measurement-gated rather than reported in full.
-    fn step(
-        &mut self,
-        input: &[u8],
-        received_at: Option<MonotonicInstant>,
-        output: &mut WriteBuf,
-    ) -> SequenceResult<Written>;
-
-    fn step_no_input(&mut self, output: &mut WriteBuf) -> SequenceResult<Written> {
-        self.step(&[], None, output)
-    }
-}
-
-ironrdp_core::assert_obj_safe!(Sequence);
 
 pub type ConnectorResult<T> = Result<T, ConnectorError>;
