+++ crates/ironrdp-rdpeudp/Cargo.toml
@@ -31,6 +31,7 @@ arbitrary = ["dep:arbitrary", "std", "bitflags/arbitrary"]
 arbitrary = { version = "1", features = ["derive"], optional = true }
 ironrdp-core = { path = "../ironrdp-core", version = "0.2", features = ["alloc"] } # public
 ironrdp-error = { path = "../ironrdp-error", version = "0.2", features = ["alloc"] } # public
+ironrdp-sequence = { path = "../ironrdp-sequence", version = "0.1", default-features = false } # public
 bitflags = "2.11" # public
 
 [lints]
