+++ crates/ironrdp-rdpeudp/src/time.rs
@@ -4,10 +4,10 @@
 //! instant on every call that can advance time, and reports when it next wants
 //! to be woken through [`RdpeudpConnection::poll_timeout`].
 //!
-//! [`MonotonicInstant`] itself lives in `ironrdp-core`, shared with
+//! [`MonotonicInstant`] itself lives in `ironrdp-sequence`, shared with
 //! `ironrdp-connector`'s `Sequence::step`, and is re-exported here so
 //! existing callers of this crate are unaffected.
 //!
 //! [`RdpeudpConnection::poll_timeout`]: crate::RdpeudpConnection::poll_timeout
 
-pub use ironrdp_core::MonotonicInstant;
+pub use ironrdp_sequence::MonotonicInstant;
