+++ crates/ironrdp-sequence/src/desktop_size.rs
@@ -1,0 +1,6 @@
+/// A negotiated remote desktop size, in pixels.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct DesktopSize {
+    pub width: u16,
+    pub height: u16,
+}
