+++ crates/ironrdp-sequence/src/state.rs
@@ -1,0 +1,32 @@
+use core::any::Any;
+use core::fmt;
+
+pub trait State: Send + fmt::Debug + 'static {
+    fn name(&self) -> &'static str;
+    fn is_terminal(&self) -> bool;
+    fn as_any(&self) -> &dyn Any;
+}
+
+ironrdp_core::assert_obj_safe!(State);
+
+pub fn state_downcast<T: State>(state: &dyn State) -> Option<&T> {
+    state.as_any().downcast_ref()
+}
+
+pub fn state_is<T: State>(state: &dyn State) -> bool {
+    state.as_any().is::<T>()
+}
+
+impl State for () {
+    fn name(&self) -> &'static str {
+        "()"
+    }
+
+    fn is_terminal(&self) -> bool {
+        true
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
