+++ crates/ironrdp-sequence/src/written.rs
@@ -1,0 +1,30 @@
+use crate::sequence_error::{SequenceError, SequenceErrorExt as _, SequenceResult};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum Written {
+    Nothing,
+    Size(core::num::NonZeroUsize),
+}
+
+impl Written {
+    #[inline]
+    pub fn from_size(value: usize) -> SequenceResult<Self> {
+        core::num::NonZeroUsize::new(value)
+            .map(Self::Size)
+            .ok_or_else(|| SequenceError::general("invalid written length (can't be zero)"))
+    }
+
+    #[inline]
+    pub fn is_nothing(self) -> bool {
+        matches!(self, Self::Nothing)
+    }
+
+    #[inline]
+    pub fn size(self) -> Option<usize> {
+        if let Self::Size(size) = self {
+            Some(size.get())
+        } else {
+            None
+        }
+    }
+}
